requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
idna==2.8
pycrypto==2.6.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
